Make live server not start up automatically
By starting up automatically the live server imposes some high costs on
tests that need it when they may not be ready yet. If we allow the user
to specify when the live server starts we can allow them to add new
routes to the server that they only need for the test. This is very
useful for people writing libraries to be used in flask rather than
writing applications using flask